+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Add `VerticalStepper.Link` styles when is an `a`.
+- Fix: Fix `WarningIcon` warning.
+
 ## [2.1.0] - 2019-04-11
 
 Features: